# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Home Appliance Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products, admin]
      summary: Create a new product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [products, admin]
      summary: Update a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products, admin]
      summary: Delete a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List current user's orders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Paginated orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place a new order from cart (cash on delivery)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderBody"
      responses:
        "201":
          description: Placed order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request (e.g. insufficient stock or empty cart)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order detail with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /upload/product-image:
    post:
      operationId: uploadProductImage
      tags: [admin]
      summary: Upload a product image (admin only)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Image URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageUploadResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        category:
          type: string
        stock:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - price
        - category
        - stock
        - createdAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        category:
          type: string
        stock:
          type: integer
      required:
        - name
        - description
        - price
        - category
        - stock
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        category:
          type: string
        stock:
          type: integer
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - limit
        - totalPages
...
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        shippingName:
          type: string
        shippingPhone:
          type: string
        shippingAddress:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - totalPrice
        - status
        - shippingName
        - shippingPhone
        - shippingAddress
        - createdAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        priceAtPurchase:
          type: number
        product:
          $ref: "#/components/schemas/Product"
      required:
        - id
        - orderId
        - productId
        - quantity
        - priceAtPurchase
    OrderWithItems:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        shippingName:
          type: string
        shippingPhone:
          type: string
        shippingAddress:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required:
        - id
        - userId
        - totalPrice
        - status
        - shippingName
        - shippingPhone
        - shippingAddress
        - createdAt
        - items
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - orders
        - total
        - page
        - limit
        - totalPages
    PlaceOrderBody:
      type: object
      properties:
        shippingName:
          type: string
        shippingPhone:
          type: string
        shippingAddress:
          type: string
      required:
        - shippingName
        - shippingPhone
        - shippingAddress
...
    ImageUploadResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url