# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Home Appliance Store API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  name: string;
  description: string;
  price: number;
  imageUrl?: string | null;
  category: string;
  stock: number;
}